refactor(binding): merge connect-queue into binding
